+++ CheeseQuest2/file/InvalidSaveNameException.java
@@ -1,0 +1,8 @@
+package file;
+
+/**
+ * Save name cannot contain invalid file name characters
+ */
+public class InvalidSaveNameException extends Exception {
+    private static final long serialVersionUID = 1L;
+}