+++ CheeseQuest2/file/InvalidSaveNumException.java
@@ -1,0 +1,8 @@
+package file;
+
+/**
+ * Save num outside save array bounds
+ */
+public class InvalidSaveNumException extends Exception {
+    private static final long serialVersionUID = 1L;
+}